@nicoddemus 2015-07-16
Closing this, reopened this for `pytest-2.7` in #851